fix init panel to allow default + 1 custom email addresses
The init panel now presents the default email contact addresses
with checkboxes for inclusion/exclusion, and also lets you add
one custom contact email to taste.

This is necessarily a bit fiddly to do, since we now have to
construct the Mailer during initialisation rather than set it
up before the GUI has been displayed.  But it's not doing
anything impossible, and it seems to work OK.