be more careful about email errors
If the startup email fails, there startup can't proceed
(this is probably because an address was entered wrong).
This does mean that potentially time-consuming work is done
on the EDT, but it's probably better than the (easy) alternative in
this case, which is that the application runs with email broken.

Email failures are now also signalled with a popup message.