use Address objects not Strings for mailer send/receive
The DmConfig now has Address-typed keys not String-typed keys
for Mailer send and receive targets.  This helps, because the
javax.mail.internet.InternetAddress parser can do parsing and
reject them with a ConfigException if they are not syntactically
valid.

At the same time, there is a fix for a bug that was setting
the recipients of a Message incorrectly (I misunderstood the
setRecipients arguments, partly because I was using a String
not Address objects).  It was only sending the message to the
last-named recipient.  It should now be doing it correctly.
My tests haven't proved this, but there may be some other
cause for that.